refactor(leases): unify Long/Short lease details behind shared core
LongLeaseDetails.vue and useShortLeaseDetails.ts shared the same reactive
skeleton (600ms-debounced Skip fee fetch with single-teardown timer, watch,
and the interest/downpayment/borrow computeds) while diverging in a handful
of localized spots. Extract that skeleton into useLeaseDetailsCore (base +
core) and keep every asymmetry in a clearly named per-position composable
rather than behind a mode conditional, mirroring the forms trio:

- fee-routing target: Short settles into the protocol stable
  (lease.total.ticker), Long swaps into the leased asset itself
- liquidation: Short calculateLiquidationShort(stable, unit), Long
  calculateLiquidation(unit, stable); percent sign per position
- sizeAmount, borrowAmount, swapFeeAmount, lpn source, and the Short-only
  totalAsset/totalLoan vs Long-only stable stay in their own files

Drains LongLeaseDetails.vue (the last script-heavy details SFC); both SFC
templates stay byte-identical. Deletes the dead swapFee ref (written, never
read) from both positions in the process.

Closes #263
